STABLE12-drop-h-lock-around-putconn-20041027
the problem with rx_PutConnection is h_FreeConnection uses the host global lock

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

the problem with rx_PutConnection is h_FreeConnection uses the host global lock
but it shouldn't matter (callback conn should be a client not server conn and
thus not h_FreeConnection on cleanup)

further, we should free the conns we GetConnection'd not the ones still in
the host struct at the end.


(cherry picked from commit 2fbfef7f9157aad571307564faae1ce7aae4e3be)