libafsrpc-assertion-failed-20041202
FIXES 16432

need AssertionFailed in libafsrpc. ndon't need casestrcpy in libafsauthent
VS: ----------------------------------------------------------------------

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 16432

need AssertionFailed in libafsrpc. ndon't need casestrcpy in libafsauthent