winnotes-20041207
update text files for StoreAnsiFilenames.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Allow users to choose to store file names in AFS using ANSI code pages
instead of OEM code pages.