disconnected-20050212
some changes i had laying around that i should flush. i should find someone to work on this for real

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

some changes i had laying around that i should flush. i should find someone to w
ork on this for real

====================

push the remainder of this stuff, including better copies of some of the files
i had. in theory this is all of disconnected afs. it won't build as-is.