windows-smb_vc_t-refcounts-20050321
conditionalize the assertions on smb_vc_t refcounts as they are only
for debugging.  We know there is a problem but don't want to deal with
it for the 1.3.80 release.

added logging info for each hold and release.  perhaps that will provide
a clue in the future

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

When copying a packet we must increment the vcp refcounts