remove-insecure-and-ntp-20050403
rcmds, inetd, ftpd and ntp are gone. leaving the files in the cvs head for now.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

rcmds, inetd, ftpd and ntp are gone. in 1.4, files also go away.