windows-rx-ts-fpq-20050418
FIXES 17805

Add a new function called rx_StartClientThread() which parallels the
behavior of rx_StartServer but is used for initializing client threads.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 17805

Add a new function called rx_StartClientThread() which parallels the
behavior of rx_StartServer but is used for initializing client threads.

Also, disable RX_TSFPQ_ENABLE for Windows as it slows the client down.