STABLE14-windows-callback-race-20050427
cm_EndCallbackGrantingCall contained a race condition due to the release
of the cm_callbackLock in the middle of the for() loop.  The race was
removed by optimizing out the call to cm_CallbackNotifyChange().  There
is no reason this needed to be called once per callback revoke in the
list.


(cherry picked from commit 8085bc4d476a2e77b26454929cdfa1d034c754f7)