STABLE14-windows-afslogon-20050428
Add new "logon" event handler in order to establish a smb connection
to the AFS client service with the true logon name as early as possible.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Add new "logon" event handler in order to establish a smb connection
to the AFS client service with the true logon name as early as possible.

Add "replace_afs.cmd" file