windows-dnlc-corruption-20050517
It has been noticed that on startup there may be corruption of the
dnlc freelist.  In particular, some items on the list are actually
in use.  This patch will cause the dnlc to be purged if corruption
of the freelist is detected.  It also makes changes to ensure that
items placed onto the freelist are in fact cleaned to prevent confusion.

The purge function is corrected to set that magic field to the correct
value.

The validation function is fixed to not CloseHandle the same handle
twice.