unroll-macos-104-20050523
we'll finish it on a branch and remerge; 1.3.83 needs to happen and 1.4 needs to go out.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

we'll finish it on a branch and remerge; 1.3.83 needs to happen and 1.4 needs to
 go out.