pr_init-fix-20050623
for the longest time aklog implementations with automatic pts registration
have included a work around for a bug in pr_Initialize.  The bug fails to
reset the connection when the cell changes.  This patch fixes that bug.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

for the longest time aklog implementations with automatic pts registration
have included a work around for a bug in pr_Initialize.  The bug fails to
reset the connection when the cell changes.  pr_Initialize has been fixed
so we can remove the hack.

====================

remove the pr_Initialize force cell change hack from afskfw lib