windows-logon-20050630
Add a method to disable the deletion of tokens at logoff

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

add a registry entry to prevent token destruction at logoff