pr_Initialize-fix-20050707
The previous patch to pr_Initialize() opened an opportunity for
afsconf_Check() to be called with a NULL pointer if tdir == 0.
Prevent this occurrance.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

fixed one null pointer reference but not the other.

====================

commit the rest of jeff's fix