* Rework how PAM modules are handled.     - Build against the pthread-aware libraries.  (Closes: #297804)     - Use the shared library objects instead of building separate PIC       objects, eliminating the need for many Makefile modifications.     - Include pam_afs.krb.so as well.  (Closes: #165972)
This is still a work in progress, but I want to get it into the repository
as a checkpoint since everything seems to work.  Remaining to be done is
the cleanup of the old Makefile rules to build the pic directory, which
should no longer be needed.

Tested with unstable ssh-krb5 and login so far, with no problems.