cache-autotune-20050817
This patch tries to implement the afsd default tuning parameters
discussed in the thread starting at
https://www.openafs.org/pipermail/openafs-devel/2005-May/012158.html

I took the liberty of adding chunksize-tuning to the memcache too,
with the motivation that people using large memcaches usually wants
better bulk performance too.

It seems to work for me using both disk cache and memcache of various
sizes.