windows-trace-log-to-dbgview-20050820
Added a new option for viewing the trace log data in real time

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Include the Thread ID in the output to make it usable for debugging
deadlocks.

====================

alter the afsd_init.log tag for the TraceOption to not be
Windows Event Log specific.