STABLE14-windows-remove-log-messages-20050826
The log message added to buf_LockedCleanAsync() during the debugging
post 1.3.8201 is output for 1/10th of all buffers once every 5 seconds.
This is a huge performance hit.  Move the message so that it is only
output for buffers that are actually dirty.

Also, change the algorithm so that the sqrt() of the number of buffers
are checked every 5 seconds instead of 1/10th.   This will do a better
job with very large cache sizes.


(cherry picked from commit ee937439adc8d8ea214c9d53aee0ee5329697c80)