STABLE14-windows-servers-all-down-do-not-retry-20050915
remove the logic in cm_Analyze that performs a flush of the server
data and retries when all of the servers are marked down (aka ALLOFFLINE).
Instead return an immediate error to the caller.  The servers will be
checked by the background daemon thread and made available when they are.


(cherry picked from commit 18bb5faec22ac8c000b7964eef635a1213500654)