STABLE14-windows-icons-20050919
Associate the OpenAFS icon with the Control Panel and Service and
the installers

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Update icon to OpenAFS logo


(cherry picked from commit cc8cfd47bc506396501c9bd2225dd907567e36b7)