windows-release-notes-20050925
Update the installed documentation to include the new OpenAFS for
Windows release notes.   No IBM documentation is being removed but
the top level index page has been re-written to indicate that the IBM
docs are out of date.   The new OAFW Release Notes have been added to
the ReleaseNotes with a link.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Fix the link to the OAFW notes.

====================

fix the links to point to .htm instead of .doc