STABLE14-windows-byte-range-locks-20050816
remove an unwanted assertion and move the resetting of scp->serverLock
to -1 into cm_LockMarkSCacheLost() so that others do not forget to set
it.  cm_LockMarkSCacheLost() is always called when the scp->mx is held
so it is ok to do so.


(cherry picked from commit 14408c7e88c198d80ce89e90e209f48685da89da)