STABLE140-windows-wix-product-id-20051005
As the file name of the OpenAFS installer changes with each release,
so must the Product ID (but not the Upgrade ID).  Therefore, do not
hard code the Product ID in the Makefile, instead auto-generate it
as part of the installer.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Remove staticly assigned product ids from the makefiles.
Product Ids will be generated on the fly by Wix.


(cherry picked from commit b9b6cb7c25cb3bcd4d34d1a52a7915c9ba46df06)