dedebug-20051007
perhaps this should never be pulled up.

anyway, implement dentry cache status dumping for linux. and provide a tool to dump it

disabled (tool, not rpc) by default

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

perhaps this should never be pulled up.

anyway, implement dentry cache status dumping for linux.^? and provide a tool to
 dump it

disabled (tool, not rpc) by default

====================

perhaps this should never be pulled up.

anyway, implement dentry cache status dumping for linux.^? and provide a tool to
 dump it

disabled (tool, not rpc) by default