macos-build-dmg-20051021
probably this script should take an argument which is the version string to use, and use it in the volume name and dmg name

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

probably this script should take an argument which is the version string to use,
 and use it in the volume name and dmg name