STABLE14-tbutc-solaris-and-vos-api-fixes-20051024
FIXES 22063

as tested for this and 22316

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 22316

the rest of this, since i forgot to apply the other file earlier
oops


(cherry picked from commit 834507571e397dc3a07550ef86ca2c159068d1ff)