windows-afsd-service-20051121
add logic to process VNOVNODE in cm_Analyze.  Force re-evaluation of symlink
strings and flush the stat cache entry.

force the use of new rx_connections when the server is marked down.

prevent server objects from being freed if user preferences are set.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

improvements to VNOVNODE fix