fix-cbd-20051213
The attached patch makes the cbd program (the stand-alone version of
the fileserver's callback.c) work again.

(Now I expect 99% "what's that?" and 0% "what a relief!" reactions. As
a reminder that's one of the things you might go hunting for after a
'kill -XCPU' to the fileserver - this program is probably used
infrequently but the older you get... )