STABLE14-viced-multiple-ports-per-client-20051208
this patch replaces the single address in the Interface structure
with an AddrPort structure.  This enables the servers to send
messages to the correct port if the port on the Interface does not
match the port stored in the client itself.


(cherry picked from commit 71f9628a437e0ac7b9b9754d71c06ca5c3408061)