windows-no-fds-20051217
when collecting rx statistics in response to an RPC query, the rx library
attempts to enumerate the number file descriptors in use.  This is fine
except that file descriptors are a C Run Time Library concept on Windows
and are not related to networking.  In Visual Studio 8, the run time library
will assert() if an invalid file descriptor is passed to fstat() which is
the test used to determine if a file descriptor is valid.

This patch simply returns 0 for the number of file descriptors in use
because that is what would have been returned anyway with the existing
code.  What we probably want to return is the number of open socket
handles.