* Call kernel-version correctly from prep-modules, which fixes build     issues with manually-built kernels.  (Closes: #344851)
Also always trust the KVERS determined by the Makefile in prep-modules
if available, since otherwise we end up with broken packages when the
makefile installs the module assuming one package name and the control
file tries to build a package with a different name.

Clean up a few remaining remnants of @sys handling when building kernel
modules.  Since the kernel module build never runs any sort of install
target, this shouldn't be needed.