STABLE14-windows-64-bit-type-safety-20051105
this is only a partial pullup
64-bit type safety changes required for successful compilation
on Windows 64-bit systems with the VS 2005 compiler

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

this is a partial pullup
64-bit type safety changes required for successful compilation
on Windows 64-bit systems with the VS 2005 compiler