STABLE14-viced-host-nat-20060207
When we discover a new address for an existing host we can take the
opportunity to cleanup the hash table and the interface list if the
new address differs from an existing address only by the port number.
In that case we know the client is communicating to us from behind a
NAT and the old addr/port number combination is no longer going to
be of use to us.


(cherry picked from commit e4c21e9c6fe4a030cf271b614d8e3d78130e4fc9)