STABLE14-html-link-fixes-20060228
Make the mentions of subcommands in the fs command introduction links to
the relevant pages, and add to README a to-do note to do this for the rest
of the introductory pages.

Fix links to man pages that contain underscores by working around a bug
in Pod::Simple.


(cherry picked from commit 6321cca1a5bdda2c9e8d9688d220eff0edd850f9)