viced-preamble-postamble-20060301
If the host obtained during the CallPremable and CallPostamble differ,
then there is the possibility of a hold leak.  This patch is a quick
attempt to deal with the potential problem.  If the problem occurs
we log it in order to know for sure if this is in fact a problem.

If it is, then at a later date we should perform a more extensive
re-write in which the CallPreamble is used to obtain the host and the
client and pass them both back to the caller which would not be
required to obtain them again at a later date.