windows-init-20060306
fix sysname string on amd64

add registry configurable CallBackPort to allow for forcing the
client to use different ports when talking with the file server.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

add registry configurable CallBackPort to allow for forcing the
client to use different ports when talking with the file server.

Add UAE errors to cm_Analyze debug output