md5-inode-numbers-20060405
FIXES 25849

an interface which lets you select md5 based inode numbers, and the csupporting code

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 25849

an interface which lets you select md5 based inode numbers, and the c^?supportin
g code