windows-query-afsid-20060420
In order to properly handle the local allocation of locks
the cache manager must enforce the locking rules associated
with PRSFS_WRITE and PRSFS_INSERT and PRSFS_LOCK.  Insert
affects "new files" which in AFS are defined as any file that
is owned by the user.  Therefore, we must know the afsid of the
user so that it can be compared to owner of the file.

This commit includes a query using PR_SNameToId where the name
is the name specified by the entity setting the token.  There
does not currently exist a Protection Service RPC to return the
AFSID of the entity that the rx connection was authenticated as.
Such a function will have to be added in order to make this
functionality secure.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Export functions missing from the UNIX version and the PR_ functions
required by the Windows Cache Manager.