STABLE14-windows-pthread-cleanup-20060424
A number of crash reports have been filed with Microsoft in afspthread.dll.
The crashes are occuring as the processes are terminating.  In order to
attempt to address this problem this patch adds support for a DllMain()
entry point that is used to cleanup Thread Local Storage and the various
RX queues.  One theory is that processes are loading and unloading DLLs
that are linked to afspthread.dll leaving the pthread data in a very
inconsistent state after each unload.  By cleaning up the static data
during the unload, if afspthread is loading again it will be forced to
initialize the data once again.


(cherry picked from commit edff1e8788a2cadfd6bb9e08f9cfdbfb491924a6)