STABLE14-viced-host-mobile-client-20060505
Do not delete an old host entry because the UUID because the UUIDs
of the clients at the old addresses are wrong.  Remove the addresses.

Do not allow the same host,addr,port combination to be added into
the hash table multiple times.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Add some more logging

Do not allow 'code' to be re-used for probing the uuid of the old host
primary address because we are relying on the value of 'code' to indicate
whether or not RCallBackConnectBack failed.  If it did, we mark the
client as VENUSDOWN, if not we marked it as RESETDONE.

====================

log the old addr, not the new one


(cherry picked from commit 69c7d8ae707fd1a145b8eec756526603ddd370b7)