STABLE14-windows-no-freelance-vldb-failover-20060622
When Freelance mode is not in use the cache manager must load the root.afs
volume of the default cell in order to be able to start.   Unfortunately,
if the first vlserver fails the request to obtain the volume information
will timeout and no retries will be performed.  This patch performs up
to 10 retries in the initialization code in order to enable the vlservers
to be marked down so that failover can occur.


(cherry picked from commit 4859232df55c94232a70d602cafe6a8bcc72d9dd)