STABLE14-windows-report-quota-diskfull-20060624
overquota and diskfull errors are detected while writing dirty buffers
to the file server.  A flag is set on the scp so that the error can
be returned to the caller.  cm_FSync was clearing the flag without
returning the error. This permitted calls to Win32 WriteFile to report
success even when they failed.


(cherry picked from commit 8e0033ba73ef9551cb24052ef90db677e4adad50)