DEVEL15-windows-bulkstat-eaccess-20060628
New CM_SCACHEFLAG_EACCESS flag is used to track whether the status
information was the result of an EACCESS error during an inline bulk
status call.  If so, we have a callback but did not get any status
info.  When this flag is set we include the entry in future bulk status
calls but otherwise treat the cm_scache entry as if it has a valid
callback.  The callback will be revoked if the ACLs on the object
changes.


(cherry picked from commit a8dc93bf9f93a1ab74972c41d09af01da93ff230)