aklog-man-page-20060731
Rewrite the aklog man page in POD and add documentation of the new flags
and .xlog.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Fix my misunderstanding of what rxkad2b is.

====================

Fix unterminated B<>.