shlib-overhaul-20060802
Overhaul how we build shared libraries.  Pull the code for building and
installing shared libraries out into simple shell scripts that we can then
evolve as we need to and make more sophisticated.  Remove the details of
how to use a mapfile from the Solaris Autoconf-driven configuration and
move them to the Solaris case of the build script.  In the process, make
it possible to build more than one shared library in a single directory.

Remove the last vestiges of libafssetpag from the sys directory; kopenafs
now replaces all of that.

Only link in syscall.o in the sys and kopenafs directories on those
platforms that actually care so that we don't have to do work to generate
a valid .o file that can be included in an archive.