DEVEL15-windows-globalautomapper-thread-20060807
Move the establishment of the GlobalAutoMapper drives to a separate
thread in order to prevent OpenAFS from reporting that the service did
not start.

OpenAFS considers the GlobalAutoMapper as deprecated functionality so
it is ok to avoid the service failure report.


(cherry picked from commit f45d2d8e67e43faf3442031e29c681d24db10824)