windows-more-misc-fixes-20061005
replace all DebugBreak() calls with osi_panic.

remove calls to cm_RecycleSCache from VNOVNODE handler as its not
lock safe to drop and re-obtain locks to adhere to the locking
hierarchy

restore reference to CM_SCACHESYNC_STOREDATA that should
not have been removed.

remove attempts to adhere to locking hierarching from cm_GetNewSCache.
dropping the cm_scacheLock is too dangerous

add a check to ensure that the scache refcount is not zero if there
is a valid smb_fid_t pointing to it.

more debug logging