DEVEL15-windows-buf-flush-20061109
FIXES 44330

The return value of buf_CleanAsync is not an error value but a flag
indicating whether or not the buffer was in fact dirty.  Do not treat
it as an error.  The error state is stored in the buffer itself.


(cherry picked from commit 398ceb0775a0663829ce4c44a8ec08d9e63bd1ca)