windows-smb-dont-crash-vista-20061128
Unlike previous versions of the OS, Vista performs a shutdown on
the Microsoft Loopback adapter just like it would on a real adapter.
This causes the smb_Listener threads to trigger a panic during a
suspend/hibernate power event.

The fix is to unbind from the network adapters in response to a
suspend/hibernate power event and then rebind to the adapters
when a resume power event is received.  Note that the resume events
are not reliably delivered so it is possible the afs service will
not be accessible.  However, this is the best we can do.