DEVEL15-rx-fix-lock-init-20061216
When rxBind was added it made an assumption that rx_GetIFInfo could be
called before rx_InitXYZ.  This is true on non-Windows platforms, but
on Windows rxGetIFInfo relies on an initialized mutex.  This patch adds
a DllMain for Windows in order to initialize the mutex object upon DLL
load.


(cherry picked from commit 5b56a0b8ebbdfffa46b90d45b06253c3c6fade15)