DEVEL15-windows-optimize-smb-dir-search-if-no-wildcard-20061217
When performing a SMB FindFirst/FindNext/FindClose operation if there
are no wildcards involved, we can optimize the case and turn it from
O(n) to O(1) where 'n' is the number of entries in the directory.
This can be done by performing a cm_Lookup() and if it succeeds,
constructing the appropriate response instead of parsing the contents
of each buffer associated with the directory looking for matches.

Without this optimization, FindFirst operations on directories containing
thousands of entries can take a large number of seconds to complete.


(cherry picked from commit a57531592bfab0b36a4d7fb58b556b848e6f75c8)