bosserver-avoid-bosconfig-truncation-20061221
why the heck would it be a good idea to rewrite the BosConfig while we are starting, exactly?

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

why the heck would it be a good idea to rewrite the BosConfig while we are start
ing, exactly?