RXK5-rxk5-20070123
heavily massaged by shadow@dementia.org to fit the head (and the src/kauth/Makefile.in changes are partly missing)

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

undo my own mistake