windows-drive-mappings-20070202
FIXES 45545

Correct the return types and values of the Window Procedures.
This fixes the drive mappings problems.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 45545

Remove "Drive " from the drive mappings.  This way the drive letters
can be selected with a single key press.