DEVEL15-windows-bkgdaemon-retry-failed-requests-20070203
background request functions such as BkgFetch and BkgStore do not
return success or failure.  the bkgDaemon thread simply executes
the request and discards it regardless of whether or not the request
was completed.  this can result in background store event failing
to be written to the file server.

modify the background request functions to return success or failure.

modify the background daemon to put failed requests back onto the queue
if the reason for failure was transient.

modify the background daemon only to process requests when the servers
for the volume are marked up.


(cherry picked from commit 37d16dd70bdc88af02db4ed0307b2ec56a4e9a34)