windows-buffer-mgmt-20070204
Instead of using a fixed sized hashtable of 1024 entries regardless
of the number of buffers, compute the hashtable to be large enough
for on average 7 buffers per bucket.

Remove the 'size' field from cm_buf_t since it is never used.

Add more info to the output of cm_buf_t objects when using "fs memdump"