multiple-local-realms-20070201
FIXES 52459

when the name is foreign, execute the body of afs_is_foreign_ticket_name

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 52459

The Name to ID RPC must distinguish between local and foreign names