DEVEL15-windows-plug-n-play-network-fix-20070307
Move the detection of which LAN adapter to use from smb_Init to
smb_NetbiosInit so that it is executed after the service is resumed
via a power management event.  Otherwise, when the network comes back
up the service attempts to bind to all LAN adapters instead of just
the loopback or the configured one.


(cherry picked from commit 5332348bdf68491670cc94ff1f9bc527876304fa)