windows-registry-20070310
move the AFS Server Manager and AFS Account Manager data cache from
the TransarcCorporation key to the OpenAFS key.  The data formats are
not compatible between the two versions and we don't want to be forced
to erase data if users switch back and forth between the two products
during OpenAFS evaluation.