DEVEL15-windows-afslogon-20070406
Fix eventlog reporting.  Do not attempt to log an event if the event
source registration fails.  Use DebugEvent0 instead of DebugEvent
when there are no parameters.

Modify the LOOKUPKEYCHAIN macro to recognize ERROR_MORE_DATA errors.

Fix the reading of Domain specific configuration for LogonScript and
TheseCells.  Previously the dwSize value was being overwritten so that
subsequent RegQueryValueEx call would fail.

Fix a memory leak in the TheseCells reading code.

Add support for Domain specific "Realm" specification.  The realm is
the realm to be appended to the username.  When logging in as a domain
or to the local machine, the specified "Domain" name is not going to be
a valid realm name.

Construct a proper principal name based upon the domain specified realm
for use in obtaining tokens with KFW.

If the domain specified "TheseCells" list includes the default cell,
do not obtain tokens twice.


(cherry picked from commit a47faf3eea14931fc5edf2d90550729dc78322df)