windows-wix-beta-upgrades-20070503
This change is not as important for OpenAFS as it was for KFW as
OpenAFS does not issue betas with the same build numbers as final
releases.  However, for users of daily builds when build numbers
are not being incremented it will provide benefit.

We modify the upgrade rules to include the current version number
and then block the uninstall if the currently installed package
is the one that is being installed.

The change applies to both the regular installer and the 32-bit
tools for WOW64 environments.