DEVEL15-windows-afsd-list-acl-only-dir-20070516
FIXES 60258

Do not return access denied when applying directory patches if the
user does not have read permission.  This is the case we want to
fake the directory entries for.  Also, make sure we set the directory
attribute on non-files so that the path can be accessed via the
Explorer Shell.


(cherry picked from commit 03e3b77553bed4d395393a53d81069c71d67dbea)