DEVEL15-windows-fake-stat-20070622
When we are faking the status data we can use the vnode value to determine
if the object should be treated as a directory or file.  even is a directory
and odd is a file.  This works even when we have never successfully
obtained status data for the object.


(cherry picked from commit afa909e66ca66db50b71434ad0396d9b5a7af936)