windows-misc-20070622
there were two sets of registry values that could be used to configure
the daemon thread check intervals.  keep the one that was documented
in the release notes and discard the other.

Add a registry value "daemonCheckOfflineVolInterval" to configure the
offline volume check interval.

Ensure that the cm_GetConn... functions initialized the output variables
to NULL on error.