DEVEL15-windows-do-not-expire-cbs-when-servers-are-down-20070802
if all of the servers are down when a callback is due to expire
delay the expiration until at least one server is available.
this prevents some applications that are running when the CM
is off the network from failing if their pages are swapped out.


(cherry picked from commit 10b76dffa0ece0cca898637a00d34b27809b9eb9)