windows-no-more-infinite-recursion-20070802
FIXES 15855

In cm_NameI, keep track of what fids were crossed when evaluating the
path.  If we discover a loop return a CM_ERROR_TOO_MANY_SYMLINKS error.

This resolves the complaint that when infinite loops are hit in the afs
name space the afsd_service.exe process uses 100% of the CPU.