windows-afsd-dirop-20070830
Do not attempt to obtain a write lock on a directory
if we already know that we aren't going to perform any
local directory updates.

Add the CM_ERROR_BPLUS_NOMATCH error which has the same
meaning as CM_ERROR_NOSUCHFILE except that it indicates
that we came to that conclusion as the result of a bplus
search.  This provides us the ability to short circuit
additional directory searches since we know the answer
is final.