DEVEL15-windows-callbacks-20070909
when suspending only clear callbacks for servers that are in an up state.
if they were in a down state, then we couldn't send the clear callbacks
message but more importantly, we need to leave the callback info current
because the callbacks should not be cleared until the server is once again
reachable.


(cherry picked from commit 7a6e406b803115176b9d8d2c235f5c729631103f)