DEVEL15-windows-version-info-20070919
FIXES 71397

Many of the OpenAFS binaries that are installed on end user machines
did not have version information.  This is problematic as it is impossible
to determine which version of a binary the crash report was filed against
when a report is received by Microsoft Windows Error Reporting service.


(cherry picked from commit feab12b929564a27a77bb84278291493003b605f)