libadmin-minimize-dns-lookups-20071003
Some cells have a large number of servers listed in the vldb that
no longer exist and/or do not have PTR records in DNS.  The algorithm
used to obtain the "server names" from IP addresses involves opening
a cell handle, creating an iterator of all servers, and then iterating
over them until a match is found for the IP address.  This is done for
every server reference.  Each time the iterator was constructed the
vldb would be queried and gethostbyaddr() would be called on every listed
ip address.

This patch stores a copy of the server list in the cell handle along with
an expiration time generated from a 10 minute TTL.  This significant
reduces but does not eliminate the number of times that the caller is
blocked on reverse dns lookups.