windows-freelance-pioctl-20071010
pioctl commands were failing with CM_ERROR_NOSUCHCELL because the
Freelance cell object was not being added to the Cell Name and ID
hash tables.

ParseIoctlPath was not consistently returning the FID of the specified
symlink or mount point object.  The was still one more case in which
CM_FLAG_FOLLOW was being specified left over from
windows-parse-ioctl-path-20070629.  Changed the use of CM_FLAG_FOLLOW
into a 'follow' variable to ease testing and permit easy migration to
adding a 'follow' parameter to the function.

RXAFS_GetVolumeStatus is always going to fail for Freelance volumes.
Fake the response data.

UpdateVolume() was performing the wrong test for Freelance volumes.
Freelance volumes should not be marked offline even though they have
no servers.