fs-uuid-20071016
FIXES 74473

Change "fs uuid" to "fs uuid -generate" where "-generate" is a
required parameter.

This command requires root access to execute it.

Unlike the Windows variant, the UNIX version does not yet support displaying
the current UUID to the user.  When this functionality is added, "-generate"
should become an optional parameter.