DEVEL15-writedcache-enforce-xdcache-writelock-20071208
LICENSE IPL10
FIXES 59136

in order that the writes to CacheItems not be addled, we need to *actually*
hold a *write* lock when we write. not even just read as the comment says.

and certainly not none as was happening in the dir package.


(cherry picked from commit c0a04c21cf36e4c8e25be9703f715f30d1bf55a6)