DEVEL15-windows-nim-afscred-identity-initialization-20080108
LICENSE MIT

One of the challenges to deploying NIM has been the configuration of
AFS token acquisition.  This patch adds a new registry key hierarchy

  HKLM\SOFTWARE\OpenAFS\Client\Realms

which is used to configure the AFS Provider for a new identity based
upon the realm of the identity.  The Realms key contains subkeys for
each realm for which configuration data is being provided

  HKLM\SOFTWARE\OpenAFS\Client\Realms\"RealmName"

"RealmName" contains one optional value, "AFSEnabled", and subkeys for
each Cell that is to be added to the configuration.  "AFSEnabled" defaults
to 0x01.

   HKLM\SOFTWARE\OpenAFS\Client\Realms\"RealmName"\"CellName"

In "CellName", there are two optional values, "Realm" and "MethodName".
If the values are not specified, the automatic realm and method determination
algorithms are used.

These values can be added to an MSI transform or deployed via Group Policy.


(cherry picked from commit 521f139a89c2f078888b8bad37a495d23723a789)